extract extension utils from _extension
